Add files via upload
I thought it would be a good idea to create functions for various aspects of the nonlinear dynamic inversion before creating a Simulink model to ensure they all work properly. This code is for the inner loop control of the UUV. The paper appears to be missing a few constants that I couldn't find. So it doesn't run yet, but thought I would give you a chance to review