Uploaded Remus Files V1
Uploaded Remus V1 files. I did not change the name of these files because I am not sure if either code calls for either function and do not want to mess that up. 
These codes were copied from the "Verification of a Six-Degree of Freedom Simulation Model
for the REMUS Autonomous Underwater Vehicle" thesis paper. I improved the formatting a bit, and fixed errors resulting from copying the older text into matlab. This mostly was in the text, - changed to _, and % changed to X, so I changed these back to the standard format.